pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "artefex"
version = "1.0.0"
description = "Neural forensic restoration - diagnose and reverse media degradation chains"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "artefex contributors" }]
keywords = [
    "image-restoration",
    "forensic",
    "degradation",
    "jpeg-artifacts",
    "neural",
    "denoising",
    "upscaling",
    "steganography",
    "forgery-detection",
    "compression",
    "image-quality",
    "ai",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/turnert2005/artefex"
Repository = "https://github.com/turnert2005/artefex"
Issues = "https://github.com/turnert2005/artefex/issues"
Changelog = "https://github.com/turnert2005/artefex/blob/main/CHANGELOG.md"
Discussions = "https://github.com/turnert2005/artefex/discussions"

[project.optional-dependencies]
neural = [
    "onnxruntime>=1.16.0",
]
web = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "python-multipart>=0.0.6",
]
video = [
    "opencv-python>=4.8.0",
]
all = [
    "artefex[neural,web,video]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
artefex = "artefex.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/artefex"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
filterwarnings = [
    "ignore:Specified provider 'CUDAExecutionProvider':UserWarning",
]

[tool.ruff]
src = ["src"]
line-length = 100
